Enabled PREDICATES as tokens in a grammar.
If a symbol in a rule is a predicate function, i.e. it is a list, which starts
with either FUNCTION (resulting from #') or LAMBDA,
FIRST-SET and FOLLOW-SET will treat it as is (as a list of objects), which makes
it storeable into the hashmap.
DEFINE-LL-1-PARSER, and others to come, shall then make it callable with
EVAL and call it with FUNCALL.

See the docstring of DEFINE-LL-1-PARSER for some concept of how it should
work then.

Also some tests were added.